Fix issues when running gearbox under Windows
Issue 1: Spaces in the path to the Python excecutable.
In this case, win32api was used for an ugly workaround.
It seems this is not needed any more for current Python
and Windows versions, so I removed the use of win32api.

Issue 2: When running with "--reload", the gearbox command
is started with the Python executable. However, if installed
as a wheel, the command is a .exe stub which cannot be run
through the Python interpreter, but must be run directly.